refactor: stat
