+++ cosid-core/src/main/java/me/ahoo/cosid/stat/converter/GroupedPrefixConverterStat.java
@@ -11,20 +11,9 @@
  * limitations under the License.
  */
 
-package me.ahoo.cosid.stat;
+package me.ahoo.cosid.stat.converter;
 
-import javax.annotation.Nullable;
+import me.ahoo.cosid.stat.Stat;
 
-public interface Stat {
-    
-    String kind();
-    
-    @Nullable
-    default Stat actual() {
-        return null;
-    }
-    
-    static Stat simple(String kind) {
-        return new SimpleStat(kind);
-    }
+record GroupedPrefixConverterStat(String kind, String delimiter, Stat actual) implements Stat {
 }
