+++ cosid-core/src/main/java/me/ahoo/cosid/stat/generator/CosIdGeneratorStat.java
@@ -11,20 +11,13 @@
  * limitations under the License.
  */
 
-package me.ahoo.cosid.stat;
+package me.ahoo.cosid.stat.generator;
 
-import javax.annotation.Nullable;
+import me.ahoo.cosid.stat.Stat;
 
-public interface Stat {
+record CosIdGeneratorStat(String kind,
+                          int machineId,
+                          long lastTimestamp,
+                          Stat converter) implements IdGeneratorStat {
     
-    String kind();
-    
-    @Nullable
-    default Stat actual() {
-        return null;
-    }
-    
-    static Stat simple(String kind) {
-        return new SimpleStat(kind);
-    }
 }
