+++ cosid-core/src/main/java/me/ahoo/cosid/stat/generator/SegmentIdStat.java
@@ -11,20 +11,18 @@
  * limitations under the License.
  */
 
-package me.ahoo.cosid.stat;
+package me.ahoo.cosid.stat.generator;
 
-import javax.annotation.Nullable;
+import me.ahoo.cosid.stat.Stat;
 
-public interface Stat {
-    
-    String kind();
-    
-    @Nullable
-    default Stat actual() {
-        return null;
-    }
-    
-    static Stat simple(String kind) {
-        return new SimpleStat(kind);
-    }
+public record SegmentIdStat(String kind,
+                            long fetchTime,
+                            long maxId,
+                            long offset,
+                            long sequence,
+                            long step,
+                            boolean isExpired,
+                            boolean isOverflow,
+                            boolean isAvailable,
+                            Stat converter) implements IdGeneratorStat {
 }
